Move broadcaster test helpers to tests
